backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
motor==3.7.1
pydantic==2.12.0
pydantic_core==2.41.1
pymongo==4.15.3
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
